requirements.txt
openai
PyYAML
Flask
Flask-CORS
sentence-transformers
faiss-cpu
numpy
tzdata
python-dotenv
dotenv
